impact2_3

Proportion of PoC with access to health services

Usage

impact2_3(datalist)

Arguments

datalist

A list with all hierarchical data frame for a survey data set. format is expected to match the Excel export synchronized from kobo to RILD and loaded with kobocruncher::kobo_data()

Value

new calculated variable - appended or not...

Details

This indicator measures access to primary health care services for persons of concern in need of health services.

Health services refers to preventive, pro-motive, curative, rehabilitative and palliative health services.

It is linked to SDG indicator 3.8.1 on the coverage of essential health services.

The standard questions for this indicator are taken from UNHCR's HAUS monitors trends in how refugees outside of camps access and utilize health services over time. Healthcare can be delivered through a combination of community level, mobile and fixed healthcare facilities. The number, type and location of each will vary by context. The standard questionnaire module will help us to understand these differences.

Standard Questions

HACC01 - HACC04

Numerator: Population who have received the asked for health services in the previous 30 days

Denominator: Total population who have asked for health services in the previous 30 days

Formula: HACC01=1 + / ((HACC03=1 & HACC04!=7,8,96) + HACC01=1)

This indicator comes from the individual dataset Calculate those who were not able to access due to reasons unrelated to asked services (when HACC04 is 7 or 8 or 96)

Examples

## data, cf example  fct_re_map()
# datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
#                                                  package = "IndicatorCalc"))
# ## Apply calculation
# datalist <- impact2_3(datalist )
# 
# ## Visualise value
# fct_plot_indic_donut(indicator = datalist[["ind"]]$impact2_3,
#                      iconunicode = "f140")