outcome4_1

Proportion of Persons of Concern who know where to access available GBV services

Usage

outcome4_1(datalist)

Arguments

datalist

A list with all hierarchical data frame for a survey data set. format is expected to match the Excel export synchronized from kobo to RILD and loaded with kobocruncher::kobo_data()

Value

new calculated variable - appended or not...

Details

The indicator is defined as the proportion of persons of concern who know where to access at least one of the four available gender-based violence services in the aftermath of a GBV incident.

This is linked to the Inter-Agency Minimum Standards for Gender-Based Violence in Emergencies Programming.

The response options will appear as below when the standardized module is used.

Definitions:

**GBV01a (**Health): "Medical treatment and health care to address the immediate and long-term physical and mental health effects of GBV. This can include initial examination and treatment, follow-up medical care, mental health care, and health-related legal services, such as preparation of documentation and provision of evidence during judicial and related processes."

GBV01b (Psycho-social/ case management): "Psychosocial care and support to assist with healing and recovery from emotional, psychological and social effects. This includes crisis care as well as longer-term emotional and practical support for the survivor and her/his family, information and advocacy, case management, and educating family members so that they can support the survivor's healing and recovery.

GBV01c (Safety/security): "Options for safety and protection for survivors and their families who are at risk of further violence and who wish to be protected. This can include safe shelters, police or community security, relocation, or in the case of children, alternative care arrangements".

GBV01d (Legal assistance): "Legal actors will clearly and honestly inform the victim/survivor of the #' procedures, limitations, pros, and cons of all existing legal options".

Standard Questions

GBV01

Numerator: Total population who indicates knowing where to access available GBV services

Denominator: Total population

Formula: ( GBV01a=1 \| GBV01b=1 \| GBV01c=1 \| GBV01d=1) / Total population

This indicator comes from main dataset based on the respondent randomly selected for individual level

Examples

## data, cf example  fct_re_map()
datalist <- kobocruncher::kobo_data( system.file("dummy_RMS_CAPI_v2_mapped.xlsx", 
                                                 package = "IndicatorCalc"))
#> Warning: Unknown or uninitialised column: `_parent_index`.
## Apply indicator function on datalist
datalist <- outcome4_1(datalist )
#> ✔ GBV01a 
#> ✔ GBV01b 
#> ✔ GBV01c 
#> ✔ GBV01d 

## Visualise value
fct_plot_indic_donut(indicator = datalist[["main"]]$outcome1_4,
                     iconunicode = "f140")   
#> Warning: Unknown or uninitialised column: `outcome1_4`.
#> No value was supplied for plotting...